+++ src/diffusers/pipelines/cosmos/pipeline_cosmos3_omni.py
@@ -707,291 +707,6 @@ class Cosmos3OmniDiffusersPipeline(DiffusionPipeline):
         )
         return tokenizer_output
 
-    def apply_timestep_embeds_to_noisy_tokens(
-        self,
-        packed_tokens: torch.Tensor,
-        packed_timestep_embeds: torch.Tensor,
-        noisy_frame_indexes: List[torch.Tensor],
-        token_shapes: list[tuple[int, ...]],
-    ) -> torch.Tensor:
-        start_noisy_index = 0
-        flattened_noisy_frame_indexes = []
-        for noisy_indexes_i, token_shape_i in zip(noisy_frame_indexes, token_shapes):
-            spatial_numel_i = math.prod(token_shape_i[1:])
-            spatial_indexes_i = torch.arange(spatial_numel_i, device=packed_tokens.device)
-            noisy_indexes_i = (noisy_indexes_i * spatial_numel_i).unsqueeze(-1).expand(-1, spatial_numel_i)
-            noisy_indexes_i = noisy_indexes_i.clone() + spatial_indexes_i + start_noisy_index
-            flattened_noisy_frame_indexes.append(noisy_indexes_i.flatten())
-            start_noisy_index += math.prod(token_shape_i)
-        flattened_noisy_frame_indexes = torch.cat(flattened_noisy_frame_indexes, dim=0)
-        flattened_noisy_frame_indexes = flattened_noisy_frame_indexes.unsqueeze(-1).expand(
-            -1,
-            packed_tokens.shape[1],
-        )
-        return packed_tokens.scatter_add(
-            dim=0,
-            index=flattened_noisy_frame_indexes,
-            src=packed_timestep_embeds,
-        )
-
-    def patchify_and_pack_latents(
-        self,
-        latent_patch_size: int,
-        latent_channel: int,
-        tokens_vision: torch.Tensor,
-        token_shapes_vision: List[Tuple[int, int, int]],
-    ) -> tuple[torch.Tensor, List[Tuple[int, int, int]]]:
-        p = latent_patch_size
-        packed_latent = []
-        original_latent_shapes = []
-        for latent, (t, h, w) in zip(tokens_vision, token_shapes_vision):
-            latent = latent.squeeze(0)  # [C,T,H,W]
-            _, t_actual, h_actual, w_actual = latent.shape
-            original_latent_shapes.append((t_actual, h_actual, w_actual))
-            h_padded = ((h_actual + p - 1) // p) * p
-            w_padded = ((w_actual + p - 1) // p) * p
-            if h_padded != h_actual or w_padded != w_actual:
-                padded = torch.zeros(
-                    (latent_channel, t_actual, h_padded, w_padded),
-                    device=latent.device,
-                    dtype=latent.dtype,
-                )
-                padded[:, :, :h_actual, :w_actual] = latent
-                latent = padded
-            h_patches = h_padded // p
-            w_patches = w_padded // p
-            latent = latent.reshape(latent_channel, t_actual, h_patches, p, w_patches, p)
-            latent = torch.einsum("cthpwq->thwpqc", latent).reshape(-1, p * p * latent_channel)
-            packed_latent.append(latent)
-        return torch.cat(packed_latent, dim=0), original_latent_shapes
-
-    def unpatchify_and_unpack_latents(
-        self,
-        latent_patch_size: int,
-        latent_channel: int,
-        packed_mse_preds: torch.Tensor,
-        token_shapes_vision: List[Tuple[int, int, int]],
-        noisy_frame_indexes_vision: list[torch.Tensor],
-        original_latent_shapes: List[Tuple[int, int, int]] | None = None,
-    ) -> list[torch.Tensor]:
-        p = latent_patch_size
-        unpatchified_latents = []
-        start_idx = 0
-        for i, (t_c, h_c, w_c) in enumerate(token_shapes_vision):
-            if original_latent_shapes is not None:
-                _, h_orig, w_orig = original_latent_shapes[i]
-                h_padded = ((h_orig + p - 1) // p) * p
-                w_padded = ((w_orig + p - 1) // p) * p
-                h_patches = h_padded // p
-                w_patches = w_padded // p
-            else:
-                h_orig, w_orig = h_c * p, w_c * p
-                h_patches, w_patches = h_c, w_c
-            noisy_frame_indexes = noisy_frame_indexes_vision[i]
-            t_n = len(noisy_frame_indexes)
-            output_tensor = torch.zeros(
-                (latent_channel, t_c, h_orig, w_orig),
-                device=packed_mse_preds.device,
-                dtype=packed_mse_preds.dtype,
-            )
-            num_patches = t_n * h_patches * w_patches
-            if num_patches > 0:
-                end_idx = start_idx + num_patches
-                latent_patches = packed_mse_preds[start_idx:end_idx]
-                latent_patches = latent_patches.reshape(t_n, h_patches, w_patches, p, p, latent_channel)
-                latent = torch.einsum("thwpqc->cthpwq", latent_patches)
-                latent = latent.reshape(latent_channel, t_n, h_patches * p, w_patches * p)
-                latent = latent[:, :, :h_orig, :w_orig]
-                output_tensor[:, noisy_frame_indexes] = latent
-                start_idx = end_idx
-            unpatchified_latents.append(output_tensor.unsqueeze(0))
-        return unpatchified_latents
-
-    def decode_vision(
-        self,
-        patch_latent_dim: int,
-        latent_patch_size: int,
-        latent_channel: int,
-        packed_seq,
-        last_hidden_state: torch.Tensor,
-        original_latent_shapes: List[Tuple[int, int, int]] | None = None,
-    ) -> list[torch.Tensor]:
-        """Decode vision predictions from last_hidden_state. Returns preds_vision list."""
-        vision = packed_seq.vision
-        has_noisy_vision = (
-            vision is not None
-            and vision.tokens is not None
-            and isinstance(vision.mse_loss_indexes, torch.Tensor)
-            and vision.mse_loss_indexes.numel() > 0
-        )
-        if not has_noisy_vision:
-            preds_vision = torch.zeros(
-                [1, patch_latent_dim], device=last_hidden_state.device, dtype=last_hidden_state.dtype
-            )
-            preds_vision = self.transformer.vae2llm(preds_vision)
-            preds_vision = self.transformer.llm2vae(preds_vision)
-            if vision is not None and vision.tokens is not None:
-                preds_vision_list = [torch.zeros_like(tok) for tok in vision.tokens]
-                preds_vision_list[0] = preds_vision_list[0] + 0.0 * preds_vision.sum()
-            else:
-                preds_vision_list = [preds_vision]
-        else:
-            assert vision is not None
-            assert isinstance(vision.mse_loss_indexes, torch.Tensor)
-            assert vision.noisy_frame_indexes is not None
-            preds_vision = self.transformer.llm2vae(last_hidden_state[vision.mse_loss_indexes])
-            preds_vision_list = self.unpatchify_and_unpack_latents(
-                latent_patch_size,
-                latent_channel,
-                preds_vision,
-                token_shapes_vision=vision.token_shapes,
-                noisy_frame_indexes_vision=vision.noisy_frame_indexes,
-                original_latent_shapes=original_latent_shapes,
-            )
-        return preds_vision_list
-
-    def _pack_sound_latents(
-        self,
-        tokens_sound: list,
-        token_shapes_sound: list,
-    ) -> torch.Tensor:
-        """Pack per-sample sound latents into a single 2-D tensor.
-
-        Args:
-            tokens_sound: List of ``[C, T]`` tensors, one per sample.
-            token_shapes_sound: List of ``(T, 1, 1)`` tuples (from packed_seq.sound).
-
-        Returns:
-            ``[total_T, C]`` packed tensor.
-        """
-        packed = []
-        for sound, shape in zip(tokens_sound, token_shapes_sound):
-            T = shape[0]
-            packed.append(sound[:, :T].permute(1, 0))  # [C, T] → [T, C]
-        return torch.cat(packed, dim=0)  # [total_T, C]
-
-    def _unpack_sound_latents(
-        self,
-        packed_preds: torch.Tensor,
-        token_shapes_sound: list,
-        noisy_frame_indexes_sound: list,
-    ) -> list:
-        """Unpack packed sound predictions back to per-sample ``[C, T]`` tensors.
-
-        Args:
-            packed_preds: ``[total_noisy_T, C]`` predictions at noisy positions.
-            token_shapes_sound: List of ``(T, 1, 1)`` tuples per sample.
-            noisy_frame_indexes_sound: List of ``[T_noisy]`` index tensors per sample.
-
-        Returns:
-            List of ``[C, T]`` tensors (zeros at conditioned positions).
-        """
-        sound_dim = self.transformer.config.sound_dim
-        unpacked = []
-        start_idx = 0
-        for shape, noisy_idxs in zip(token_shapes_sound, noisy_frame_indexes_sound):
-            T = shape[0]
-            output = torch.zeros(
-                (sound_dim, T),
-                device=packed_preds.device,
-                dtype=packed_preds.dtype,
-            )
-            t_n = len(noisy_idxs)
-            if t_n > 0:
-                output[:, noisy_idxs] = packed_preds[start_idx : start_idx + t_n].T
-                start_idx += t_n
-            unpacked.append(output)
-        return unpacked
-
-    def encode_sound_tokens(
-        self,
-        timestep_scale: float,
-        packed_seq,
-        hidden_states: torch.Tensor,
-        target_dtype: torch.dtype,
-    ) -> None:
-        """Project sound tokens into ``hidden_states`` in-place.
-
-        Projects sound latents into ``hidden_states`` via ``sound2llm``, modality embed, and timestep embeddings.
-
-        Steps:
-        1. Pack latents: list of ``[C, T]`` → ``[total_T, C]``
-        2. Project: ``sound2llm`` + ``sound_modality_embed``
-        3. Add timestep embeddings to noisy frames
-        4. Scatter into ``hidden_states`` at ``sound.sequence_indexes``
-        """
-        if packed_seq.sound is None or packed_seq.sound.tokens is None:
-            return
-
-        sound = packed_seq.sound
-        assert sound.token_shapes is not None
-        assert isinstance(sound.sequence_indexes, torch.Tensor)
-        assert isinstance(sound.timesteps, torch.Tensor)
-        assert isinstance(sound.mse_loss_indexes, torch.Tensor)
-
-        packed_tokens_sound = self._pack_sound_latents(sound.tokens, sound.token_shapes)
-        packed_tokens_sound = packed_tokens_sound.to(target_dtype)
-
-        packed_tokens_sound = self.transformer.sound2llm(packed_tokens_sound) + self.transformer.sound_modality_embed
-
-        if sound.mse_loss_indexes.numel() > 0:
-            timesteps_sound = sound.timesteps * timestep_scale
-            with torch.autocast("cuda", enabled=True, dtype=torch.float32):
-                packed_timestep_embeds_sound = self.transformer.time_embedder(
-                    self.transformer.time_proj(timesteps_sound)
-                )
-            packed_timestep_embeds_sound = packed_timestep_embeds_sound.to(target_dtype)
-            packed_tokens_sound = self.apply_timestep_embeds_to_noisy_tokens(
-                packed_tokens=packed_tokens_sound,
-                packed_timestep_embeds=packed_timestep_embeds_sound,
-                noisy_frame_indexes=sound.noisy_frame_indexes,
-                token_shapes=sound.token_shapes,
-            )
-
-        hidden_states[sound.sequence_indexes] = packed_tokens_sound
-
-    def decode_sound_tokens(
-        self,
-        packed_seq,
-        last_hidden_state: torch.Tensor,
-    ) -> list:
-        """Decode sound predictions from transformer hidden states.
-
-        Extracts sound predictions from hidden states via ``llm2sound`` and unpacks them back to latent shape.
-        Includes a dummy forward path for graph-consistency when no noisy tokens.
-
-        Returns:
-            List of ``[C, T]`` tensors, one per sample.
-        """
-        sound = packed_seq.sound
-        has_noisy_sound = (
-            sound is not None
-            and sound.tokens is not None
-            and isinstance(sound.mse_loss_indexes, torch.Tensor)
-            and sound.mse_loss_indexes.numel() > 0
-        )
-
-        if not has_noisy_sound:
-            sound_dim = self.transformer.config.sound_dim
-            dummy = torch.zeros(
-                [1, sound_dim],
-                device=last_hidden_state.device,
-                dtype=last_hidden_state.dtype,
-            )
-            dummy = self.transformer.sound2llm(dummy) + self.transformer.sound_modality_embed
-            dummy = self.transformer.llm2sound(dummy)
-            if sound is not None and sound.tokens is not None:
-                preds = [torch.zeros_like(tok) for tok in sound.tokens]
-                preds[0] = preds[0] + 0.0 * dummy.sum()
-            else:
-                preds = [dummy]
-            return preds
-
-        assert sound is not None
-        assert isinstance(sound.mse_loss_indexes, torch.Tensor)
-        preds_packed = self.transformer.llm2sound(last_hidden_state[sound.mse_loss_indexes])
-        return self._unpack_sound_latents(preds_packed, sound.token_shapes, sound.noisy_frame_indexes)
-
     def decode_sound(self, latent: torch.Tensor) -> torch.Tensor:
         """Decode a sound latent ``[C, T]`` to a waveform ``[audio_ch, N]``.
 
@@ -1138,59 +853,6 @@ class Cosmos3OmniDiffusersPipeline(DiffusionPipeline):
             x0_tokens_sound,
             fps_sound,
         )
-
-    def encode_text(
-        self,
-        hidden_size: int,
-        packed_seq,
-    ) -> tuple[torch.Tensor, torch.dtype]:
-        """Embed text tokens. Returns (hidden_states [N_total, H], target_dtype)."""
-        packed_text_embedding = self.transformer.model.embed_tokens(packed_seq.text_ids)
-        hidden_states = packed_text_embedding.new_zeros(size=(packed_seq.sequence_length, hidden_size))
-        hidden_states[packed_seq.text_indexes] = packed_text_embedding
-        return hidden_states, packed_text_embedding.dtype
-
-    def encode_vision(
-        self,
-        timestep_scale: float,
-        latent_patch_size: int,
-        latent_channel: int,
-        packed_seq,
-        hidden_states: torch.Tensor,
-        target_dtype: torch.dtype,
-        fps: Optional[torch.Tensor] = None,
-    ) -> List[Tuple[int, int, int]] | None:
-        """Project vision tokens into hidden_states in-place. Returns original_latent_shapes."""
-        if packed_seq.vision is None or packed_seq.vision.tokens is None:
-            return None
-        vision = packed_seq.vision
-        assert vision.tokens is not None
-        assert vision.token_shapes is not None
-        assert isinstance(vision.sequence_indexes, torch.Tensor)
-        assert isinstance(vision.timesteps, torch.Tensor)
-        assert isinstance(vision.mse_loss_indexes, torch.Tensor)
-
-        packed_tokens_vision, original_latent_shapes = self.patchify_and_pack_latents(
-            latent_patch_size, latent_channel, vision.tokens, vision.token_shapes
-        )
-        packed_tokens_vision = self.transformer.vae2llm(packed_tokens_vision)
-
-        if vision.mse_loss_indexes.numel() > 0:
-            timesteps_vision = vision.timesteps * timestep_scale
-            with torch.autocast("cuda", enabled=True, dtype=torch.float32):
-                packed_timestep_embeds_vision = self.transformer.time_embedder(
-                    self.transformer.time_proj(timesteps_vision)
-                )
-            packed_timestep_embeds_vision = packed_timestep_embeds_vision.to(target_dtype)
-            packed_tokens_vision = self.apply_timestep_embeds_to_noisy_tokens(
-                packed_tokens=packed_tokens_vision,
-                packed_timestep_embeds=packed_timestep_embeds_vision,
-                noisy_frame_indexes=vision.noisy_frame_indexes,
-                token_shapes=vision.token_shapes,
-            )
-
-        hidden_states[vision.sequence_indexes] = packed_tokens_vision
-        return original_latent_shapes
 
     def check_inputs(
         self,
@@ -1410,12 +1072,8 @@ class Cosmos3OmniDiffusersPipeline(DiffusionPipeline):
 
         # Hoist per-step constants out of the loop
         config = self.transformer.config
-        hidden_size = config.hidden_size
-        latent_patch_size = config.latent_patch_size
-        latent_channel = config.latent_channel
-        patch_latent_dim = config.patch_latent_dim
-        timestep_scale = config.timestep_scale
         assert config.use_moe
+        latent_patch_size = config.latent_patch_size
         include_eog = config.joint_attn_implementation == "flex"
         has_sound = x0_tokens_sound is not None
 
@@ -1470,24 +1128,7 @@ class Cosmos3OmniDiffusersPipeline(DiffusionPipeline):
                     packed_seq.vision.tokens = [x.to(device=device, dtype=dtype) for x in noise_x_vision]
                 if noise_x_sound is not None and packed_seq.sound is not None:
                     packed_seq.sound.tokens = [x.to(device=device, dtype=dtype) for x in noise_x_sound]
-
-                hidden_states, target_dtype = self.encode_text(hidden_size, packed_seq)
-                original_latent_shapes = self.encode_vision(
-                    timestep_scale, latent_patch_size, latent_channel, packed_seq, hidden_states, target_dtype, fps=fps_vision
-                )
-                if noise_x_sound is not None:
-                    self.encode_sound_tokens(timestep_scale, packed_seq, hidden_states, target_dtype)
-
-                und_len = packed_seq.und_len
-                und_out, gen_out = self.transformer(
-                    hidden_states[:und_len], hidden_states[und_len:], position_ids=packed_seq.position_ids
-                )
-                last_hidden_state = torch.cat([und_out, gen_out], dim=0)
-
-                preds_vision = self.decode_vision(
-                    patch_latent_dim, latent_patch_size, latent_channel, packed_seq, last_hidden_state, original_latent_shapes
-                )
-                preds_sound = self.decode_sound_tokens(packed_seq, last_hidden_state) if noise_x_sound is not None else None
+                preds_vision, preds_sound = self.transformer(packed_seq)
 
                 velocity_vision = [
                     pred * (1.0 - m).to(dtype=pred.dtype, device=pred.device)
@@ -1528,24 +1169,7 @@ class Cosmos3OmniDiffusersPipeline(DiffusionPipeline):
                     packed_seq.vision.tokens = [x.to(device=device, dtype=dtype) for x in noise_x_vision]
                 if noise_x_sound is not None and packed_seq.sound is not None:
                     packed_seq.sound.tokens = [x.to(device=device, dtype=dtype) for x in noise_x_sound]
-
-                hidden_states, target_dtype = self.encode_text(hidden_size, packed_seq)
-                original_latent_shapes = self.encode_vision(
-                    timestep_scale, latent_patch_size, latent_channel, packed_seq, hidden_states, target_dtype, fps=fps_vision
-                )
-                if noise_x_sound is not None:
-                    self.encode_sound_tokens(timestep_scale, packed_seq, hidden_states, target_dtype)
-
-                und_len = packed_seq.und_len
-                und_out, gen_out = self.transformer(
-                    hidden_states[:und_len], hidden_states[und_len:], position_ids=packed_seq.position_ids
-                )
-                last_hidden_state = torch.cat([und_out, gen_out], dim=0)
-
-                preds_vision = self.decode_vision(
-                    patch_latent_dim, latent_patch_size, latent_channel, packed_seq, last_hidden_state, original_latent_shapes
-                )
-                preds_sound = self.decode_sound_tokens(packed_seq, last_hidden_state) if noise_x_sound is not None else None
+                preds_vision, preds_sound = self.transformer(packed_seq)
 
                 velocity_vision = [
                     pred * (1.0 - m).to(dtype=pred.dtype, device=pred.device)
